# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: MetaRGB
    description: REST API for MetaRGB microservices. Laravel-compatible JSON responses served via Kong → grpc-gateway → gRPC services.
    title: MetaRGB API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: grpc-gateway direct (dev)
      url: http://localhost:8080
...
paths:
...
    /api/levels:
        get:
            operationId: get_levels
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: List all levels
            tags:
                - Levels
    /api/levels/{id}:
        get:
            operationId: get_levels_id
            parameters:
                - description: URL slug
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: Get level details
            tags:
                - Levels
    /api/levels/{slug}/gem:
        get:
            operationId: get_levels_slug_gem
            parameters:
                - description: URL slug
                  in: path
                  name: slug
                  required: true
                  schema:
                    example: example-slug
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetLevelGem
            tags:
                - Levels
    /api/levels/{slug}/general-info:
        get:
            operationId: get_levels_slug_general-info
            parameters:
                - description: URL slug
                  in: path
                  name: slug
                  required: true
                  schema:
                    example: example-slug
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetLevelGeneralInfo
            tags:
                - Levels
    /api/levels/{slug}/gift:
        get:
            operationId: get_levels_slug_gift
            parameters:
                - description: URL slug
                  in: path
                  name: slug
                  required: true
                  schema:
                    example: example-slug
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetLevelGift
            tags:
                - Levels
    /api/levels/{slug}/licenses:
        get:
            operationId: get_levels_slug_licenses
            parameters:
                - description: URL slug
                  in: path
                  name: slug
                  required: true
                  schema:
                    example: example-slug
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetLevelLicenses
            tags:
                - Levels
    /api/levels/{slug}/prize:
        get:
            operationId: get_levels_slug_prize
            parameters:
                - description: URL slug
                  in: path
                  name: slug
                  required: true
                  schema:
                    example: example-slug
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetLevelPrize
            tags:
                - Levels
...
    /api/users/{user}/levels:
        get:
            operationId: get_users_user_levels
            parameters:
                - description: User ID or citizen code
                  in: path
                  name: user
                  required: true
                  schema:
                    example: user-code
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: GetUserLevels
            tags:
                - Users
...
components:
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Laravel Sanctum-style bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http